Rename output and source directories in Cloudflare Pages deployment script
The 'main.tf' script has been updated to change the output and source directories for the Archive file from 'build_folder' to the more descriptive 'app_folder' for Cloudflare Pages deployment. In 'variables.tf', an 'app_folder' variable has been added to accommodate this change and to allow flexibility in defining the folder where the app is stored.